Silence conversion messages

+++ convert_wav_to_mp3.ps1
@@ -20,12 +20,12 @@ $folders = Get-ChildItem -Path $work_dir -Directory		# store every folder in a l
 
 foreach ($folder in $folders)
 {	
-	"Getting .wavs..."
+	"Getting .wavs inside $folder..."
 	$wavs_all = Get-ChildItem -Path $folder -File "*.wav"	# store every .wav in the current folder
-	$wavs
 	
 	if ($wavs_all.Count -lt 2)					# skip processing in case there are less than two elements
 	{
+		"Too less elements in $folder. No conversion will be done."
 		continue
 	}
 	else
@@ -74,18 +74,25 @@ foreach ($folder in $folders)
 		$in = ".\$($folder)\$($wav)"								# .\2023.08.27\001.wav
 		$out = "$($output_path_archive)$($filename_archive).mp3"	# .\2023.08.27\mp3\20230827_1.mp3
 																	# converted files are stored in a folder "archiv" inside every day
-		
-		ffmpeg -i $in -vn -ar 44100 -ac 2 -b:a 192k $out			# basic conversion
+
+		if (!(Test-Path -Path $out)){
+			"Converting $in to $out..."
+			ffmpeg -i $in -vn -ar 44100 -ac 2 -b:a 192k $out > $null 2>&1	# basic conversion
+		}
 
 		# assuming that the predigt is the second element, copy it to the respective folder
 		if ($i -eq 2) 
 		{
 			$out_predigt = "$($output_path_predigt)$($filename_predigt).mp3"
-			Copy-Item -Path $out -Destination $out_predigt						# Copy the file and rename it
+			if (!(Test-Path -Path $out_predigt))
+			{
+				"Copy $out to $out_predigt..."
+				Copy-Item -Path $out -Destination $out_predigt		# Copy the file and rename it
+			}
 		}
 
 		$i++														# increase counter
 	}
-	"Finished converting."
+	"Finished converting for $folder."
 }
